Create Again Sum_2
This task is simple. You will be given the balance of n accounts. You have to find the sum of the balances where each balance is in range [L, R].

Input

The first line of the input contains n, the total number of bank accounts. 

The next line contains n integers the account balances.

The next line contains an integer q, which denotes the total number of queries to be asked next.

The next q lines each contains two integers L and R, meaning that you have to output the sum of the balances where each balance is between L and R.

1 <= n <= 105

1 <= balance <= 104

1 <= q <= 105

1 <= L <= R <= 105

Output

The q line of output contains the sum of the balances as mentioned above. 

Example

Input

4

1 2 3 4  

3

4 5

2 6

5 6

Output

4

9

0